requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
idna==3.4
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
urllib3==2.0.7
